compact_dazibao implemented (untested)

+++ src/dazibao.h
@@ -10,6 +10,7 @@
 #include <unistd.h>
 #include <stdlib.h>
 #include "tlvs.h"
+#include "utils.h"
 
 
 #define PANIC(str) {					\
@@ -41,23 +42,24 @@ int open_dazibao(struct dazibao* d, char* path, int flags);
 int close_dazibao(struct dazibao* d);
 
 /*  */
-int read_tlv(struct dazibao* d, struct tlv* buf, int offset);
+int read_tlv(struct dazibao* d, struct tlv* buf, off_t offset);
 
 /*  */
 int next_tlv(struct dazibao* d, struct tlv* buf);
 
 /*  */
-int tlv_at(struct dazibao* d, struct tlv* buf, int offset);
+int tlv_at(struct dazibao* d, struct tlv* buf, off_t offset);
 
 /*  */
 int add_tlv(struct dazibao* d, struct tlv* buf);
 
 /*  */
-int rm_tlv(struct dazibao* d, int offset);
+int rm_tlv(struct dazibao* d, off_t offset);
 
 /*
  * Compact a Dazibao file. The file must have been opened in read/write mode,
- * and the Dazibao is NOT closed by the function.
+ * and the Dazibao is NOT closed by the function. Also, the dazibao offset is
+ * NOT preserved.
  * The function returns the number of bytes saved by the compacting operation,
  * or -1 if an error occured.
  */
